custom-elements/registering-custom-elements/definition-construction-algorithm-test-004.html

<!DOCTYPE html>
<html>
<head>
<title>For SVG prototype namespace is SVG namespace</title>
<meta name="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<meta name="assert" content="If PROTOTYPE's interface inherits from SVGElement, set NAMESPACE to SVG Namespace">
<link rel="help" href="http://www.w3.org/TR/custom-elements/#registering-custom-elements">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src='../testcommon.js'></script>
<link rel="stylesheet" href="/resources/testharness.css">
</head>
<body>
<div id="log"></div>
<script type="text/javascript">

test(function() {
    var doc = newHTMLDocument();

    var svg = doc.createElementNS(SVG_NAMESPACE, 'x-foo');
    var proto = Object.create(svg.constructor.prototype);
    
    var XFoo = doc.register('x-foo-1', {prototype: proto});
    
    var xfoo = new XFoo();
    
    assert_equals(xfoo.namespaceURI, SVG_NAMESPACE, 'Wrong namespace');
    
}, 'For SVG prototype namespace is SVG namespace');

</script>
</body>
</html>
